ref(onboarding): Split django onboarding docs
