+++ static/app/gettingStartedDocs/python/django/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './django';
+import docs from '.';
 
 describe('django onboarding docs', () => {
   it('renders doc correctly', () => {
